Abstract MaskedTextField from PhoneNumberField
